Trait mw_dataformat::asm::Assembly

pub trait Assembly: Sized {
    type DisasmError: Error;
    type AsmError: Error;

    // Required methods
    fn disasm_one(
        buffer: &[Self],
        fmt: &mut Formatter<'_>
    ) -> Result<usize, Self::DisasmError>;
    fn asm_one(
        source: &str,
        buffer: &mut [Self]
    ) -> Result<usize, Self::AsmError>;

    // Provided methods
    fn disasm_all(
        buffer: &[Self],
        fmt: &mut Formatter<'_>
    ) -> (usize, Result<(), (usize, Self::DisasmError)>) { ... }
    fn asm_all(
        source: &str,
        buffer: &mut [Self]
    ) -> Result<usize, Self::AsmError> { ... }
}

Required Associated Types

type DisasmError: Error

type AsmError: Error

Required Methods

fn disasm_one( buffer: &[Self], fmt: &mut Formatter<'_> ) -> Result<usize, Self::DisasmError>

Attempt to disassemble one instruction

buffer may contain data that represents any (possibly unknown) number of instructions. This method will interpret the first, and return how many elements in the buffer were processed. The returned usize represents the offset at which the next instruction is supposed to start. Advance the slice to that offset and call this method again, to disassemble the next instruction.

fn asm_one(source: &str, buffer: &mut [Self]) -> Result<usize, Self::AsmError>

Attempt to assemble one instruction

source is the source code text string. If it contains more than one instruction, parsing will stop after the first. Given that our format is newline-delimited, you should probably only call this method on individual lines.

On success, returns the number of elements written into buffer. You should advance by that offset before any future write to the buffer, to not clobber the data written by this method.

Provided Methods

fn disasm_all( buffer: &[Self], fmt: &mut Formatter<'_> ) -> (usize, Result<(), (usize, Self::DisasmError)>)

Attempt to disassemble an entire sequence of instructions

buffer may contain data that represents any (possibly unknown) number of instructions. This method will attempt to process all of it. It will keep going until it either disassembles everything successfully, or encounters an error.

Returns the total number of instructions that have been disassembled, as well as a Result indicating whether that is all of them (the entire buffer has been processed), or disassembly stopped at some point due to an error. If there is an error, the offset within the buffer where the error occured is returned alongside the error.

fn asm_all(source: &str, buffer: &mut [Self]) -> Result<usize, Self::AsmError>

Attempt to assemble an entire sequence of instructions

source is the source code text string. If it contains more than one instruction, it will keep going and process as much as possible. It is expected to be newline-delimited, where each line represents one instruction (or is blank).

On success, returns the number of elements written into buffer. You should advance by that offset before any future write to the buffer, to not clobber the data written by this method.